+++ docs/changes.md
@@ -3,6 +3,7 @@
   * 重构
   * 增加navigator\_handle\r_awtk
   * 增加model事件
+  * 增加demo7/demo9
 
 * 2019/02/07
   * 重构